// src/app/models/user.model.interface.ts
export interface Root {
  page: number
  per_page: number
  total: number
  total_pages: number
  data: Data[]
}

export interface IUser {
  data: Data
}

export interface Data {
  id: number
  email: string
  first_name: string
  last_name: string
  avatar: string
}

export interface Support {
  url: string
  text: string
}